fix sonarqube raised issues refactor policies
